fix: compile complete

+++ pt-client/src/modules/timeline.rs
@@ -20,8 +20,8 @@ pub fn read(doc: Element) -> TimelineState {
     let mut state = TimelineState {
 
         tempo: *params.get("bpm").unwrap_or(&127) as u16,
-        meter_beat: *params.get("meter_beat").unwrap_or(&4) as usize,
-        meter_note: *params.get("meter_note").unwrap_or(&4) as usize,
+        meter_beat: *params.get("meter_beat").unwrap_or(&4) as u16,
+        meter_note: *params.get("meter_note").unwrap_or(&4) as u16,
         seq_in: *marks.get("seq_in").unwrap_or(&0),
         seq_out: *marks.get("seq_out").unwrap_or(&48000),
         loop_in: *marks.get("loop_in").unwrap_or(&0),
